saas/requirements.txt
fastapi
uvicorn
google-generativeai
fastapi-clerk-auth